ADAPT-3834: Manual story card (#273)
* ADAPT-3834: Manual story card

* Lint fix

* ADAPT-3834: Fix accidentally added word

* ADAPT-3834: Use existing component

* ADAPT-3834: Creeatee ereeusable card componeent

* Lint fix
