+++ src/components/components.js
@@ -52,6 +52,7 @@ import Section from './layout/section';
 import SimpleImage from './media/simpleImage';
 import Story from './content-types/story/story';
 import StoryCard from './cards/storyCard';
+import SaaStoryCard from './cards/saaStoryCard';
 import { SBTripCard } from './storyblok/tripCard';
 import { SBTsContentPicker } from './storyblok/tsContentPicker';
 import { SBTsContentTemplate } from './storyblok/tsContentTemplate';
@@ -117,6 +118,7 @@ const ComponentList = {
   simpleImage: SimpleImage,
   story: Story,
   storyCard: StoryCard,
+  saaStoryCard: SaaStoryCard,
   trip: TripPage,
   tripCard: SBTripCard,
   tripFilterPage: TripFilterPage,
